Create README.md
COPIED FROM https://github.com/Skylortrexler/Sentimentality-2
You might want to change it as IDK if all the features are the same. Mention about the copper and amethyst tools should be added.
Also maybe move the CF link to the "About" section (this can't be done in a pr AFAIK).